+++ packages/commonwealth/client/scripts/views/pages/proposals.tsx
@@ -14,7 +14,7 @@ import Compound from 'controllers/chain/ethereum/compound/adapter';
 import Aave from 'controllers/chain/ethereum/aave/adapter';
 import Sublayout from 'views/sublayout';
 import { PageLoading } from 'views/pages/loading';
-import { ProposalCard } from 'views/components/proposal_card/proposal_card';
+import { ProposalCard } from 'views/components/proposal_card';
 import { loadSubstrateModules } from 'views/components/load_substrate_modules';
 import { PageNotFound } from 'views/pages/404';
 import ErrorPage from 'views/pages/error';
