+++ packages/commonwealth/client/scripts/views/pages/referenda.tsx
@@ -10,7 +10,7 @@ import { ChainBase } from 'common-common/src/types';
 import Substrate from 'controllers/chain/substrate/main';
 import Sublayout from 'views/sublayout';
 import { PageLoading } from 'views/pages/loading';
-import { ProposalCard } from 'views/components/proposal_card/proposal_card';
+import { ProposalCard } from 'views/components/proposal_card';
 import { CountdownUntilBlock } from 'views/components/countdown';
 import { loadSubstrateModules } from 'views/components/load_substrate_modules';
 import ErrorPage from './error';
