+++ packages/commonwealth/client/scripts/views/pages/tips.tsx
@@ -12,7 +12,7 @@ import Substrate from 'controllers/chain/substrate/main';
 import { SubstrateTreasuryTip } from 'controllers/chain/substrate/treasury_tip';
 import Sublayout from 'views/sublayout';
 import { PageLoading } from 'views/pages/loading';
-import { ProposalCard } from 'views/components/proposal_card/proposal_card';
+import { ProposalCard } from 'views/components/proposal_card';
 import { loadSubstrateModules } from 'views/components/load_substrate_modules';
 import ErrorPage from './error';
 import User from '../components/widgets/user';
