+++ packages/commonwealth/client/scripts/views/pages/treasury.tsx
@@ -10,7 +10,7 @@ import { ChainBase } from 'common-common/src/types';
 import Substrate from 'controllers/chain/substrate/main';
 import Sublayout from 'views/sublayout';
 import { PageLoading } from 'views/pages/loading';
-import { ProposalCard } from 'views/components/proposal_card/proposal_card';
+import { ProposalCard } from 'views/components/proposal_card';
 import { CountdownUntilBlock } from 'views/components/countdown';
 import ErrorPage from 'views/pages/error';
 import { loadSubstrateModules } from 'views/components/load_substrate_modules';
