+++ packages/commonwealth/client/styles/components/proposal_card/index.scss
@@ -1,9 +1,9 @@
 @import 'shared';
 
-.ProposalCard {
+.ProposalCard.Card {
   display: flex;
   flex-direction: column;
-  padding: 24px;
+  gap: 8px;
   width: 260px;
 
   .proposal-card-metadata {
@@ -24,22 +24,6 @@
     color: $neutral-600;
     font-size: 14px;
     margin-bottom: 8px;
-  }
-
-  .cui-tag {
-    background: $neutral-50;
-    border-color: transparent;
-    border-radius: 4px;
-    color: $black;
-    font-weight: 500;
-    height: 32px;
-    margin-bottom: 4px;
-    width: fit-content;
-  }
-
-  .cui-tag.proposal-became-tag {
-    background: $neutral-200;
-    border-color: $neutral-200;
   }
 
   .proposal-title {
