+++ packages/commonwealth/client/styles/components/proposal_card/proposal_tag.scss
@@ -1,0 +1,12 @@
+@import 'shared';
+
+.ProposalTag {
+  align-items: center;
+  background-color: $neutral-100;
+  border-radius: $border-radius-corners;
+  display: flex;
+  height: 32px;
+  justify-content: center;
+  padding: 8px;
+  width: fit-content;
+}
