+++ packages/commonwealth/client/styles/pages/view_proposal/aave_summary.scss
@@ -1,4 +1,7 @@
 @import 'shared';
 
 .AaveSummary {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
 }
